requirements.txt
json
discord.py
google-api-python-client
yt-dlp
pynacl
asyncio
brew install ffmpeg
